requirements.txt
acres==0.3.0
aiofiles==24.1.0
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
ci-info==0.3.0
click==8.1.8
configobj==5.0.9
configparser==7.2.0
easyocr==1.7.2
etelemetry==0.3.1
filelock==3.18.0
fitz==0.0.1.dev2
frontend==0.0.3
fsspec==2024.6.1
h11==0.14.0
httpcore==1.0.8
httplib2==0.22.0
httpx==0.28.1
idna==3.10
imageio==2.37.0
isodate==0.6.1
itsdangerous==2.2.0
Jinja2==3.1.4
lazy_loader==0.4
looseversion==1.3.0
lxml==5.3.2
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.4.2
nibabel==5.3.2
ninja==1.11.1.4
nipype==1.10.0
numpy==2.2.4
nvidia-cublas-cu12==12.8.3.14
nvidia-cuda-cupti-cu12==12.8.57
nvidia-cuda-nvrtc-cu12==12.8.61
nvidia-cuda-runtime-cu12==12.8.57
nvidia-cudnn-cu12==9.7.1.26
nvidia-cufft-cu12==11.3.3.41
nvidia-cufile-cu12==1.13.0.11
nvidia-curand-cu12==10.3.9.55
nvidia-cusolver-cu12==11.7.2.55
nvidia-cusparse-cu12==12.5.7.53
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.8.61
nvidia-nvtx-cu12==12.8.55
ollama==0.4.8
opencv-python-headless==4.11.0.86
packaging==24.2
pandas==2.2.3
pathlib==1.0.1
pillow==11.2.1
prov==2.0.1
puremagic==1.28
pyclipper==1.3.0.post6
pydantic==2.11.3
pydantic_core==2.33.1
pydot==3.0.4
PyMuPDF==1.25.5
pyparsing==3.2.3
pytesseract==0.3.13
python-bidi==0.6.6
python-dateutil==2.9.0.post0
pytils==0.4.3
pytz==2025.2
pyxnat==1.6.3
PyYAML==6.0.2
rdflib==6.3.2
requests==2.32.3
scikit-image==0.25.2
scipy==1.15.2
setuptools==70.2.0
shapely==2.1.0
simplejson==3.20.1
six==1.17.0
sniffio==1.3.1
starlette==0.46.2
sympy==1.13.3
tifffile==2025.3.30
tools==0.1.9
torch==2.7.0+cu128
torchaudio==2.7.0+cu128
torchvision==0.22.0+cu128
traits==7.0.2
triton==3.3.0
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.1
